pyproject.toml
[project]
name = "customserviceagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.0",
]
